refactor(api): move cancel certification-center-invitation route
