+++ api/lib/application/certification-center-invitations/index.js
@@ -5,40 +5,7 @@ import { identifiersType } from '../../../src/shared/domain/types/identifiers-ty
 import { certificationCenterInvitationController } from './certification-center-invitation-controller.js';
 
 const register = async function (server) {
-  const adminRoutes = [
-    {
-      method: 'DELETE',
-      path: '/api/admin/certification-center-invitations/{certificationCenterInvitationId}',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-            assign: 'hasAuthorizationToAccessAdminScope',
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            certificationCenterInvitationId: identifiersType.certificationCenterInvitationId,
-          }),
-        },
-        handler: certificationCenterInvitationController.cancelCertificationCenterInvitation,
-        tags: ['api', 'admin', 'invitations', 'cancel'],
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            "- Elle permet d'annuler une invitation envoyée mais non acceptée encore.",
-        ],
-      },
-    },
-  ];
-
   server.route([
-    ...adminRoutes,
     {
       method: 'POST',
       path: '/api/certification-center-invitations/{id}/accept',
