+++ api/tests/team/acceptance/application/certification-center-invitation/certification-center-invitation.admin.route.test.js
@@ -64,4 +64,29 @@ describe('Acceptance | Team | Application | Route | Admin | Certification Center
       });
     });
   });
+
+  describe('DELETE /api/admin/certification-centers/{id}/invitations/{certificationCenterInvitationId}', function () {
+    it('should return 204 HTTP status code', async function () {
+      // given
+      const adminMember = await insertUserWithRoleSuperAdmin();
+      const certificationCenterInvitation = databaseBuilder.factory.buildCertificationCenterInvitation({
+        certificationCenterId: databaseBuilder.factory.buildCertificationCenter().id,
+        status: CertificationCenterInvitation.StatusType.PENDING,
+      });
+
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject({
+        method: 'DELETE',
+        url: `/api/admin/certification-center-invitations/${certificationCenterInvitation.id}`,
+        headers: {
+          authorization: generateValidRequestAuthorizationHeader(adminMember.id),
+        },
+      });
+
+      // then
+      expect(response.statusCode).to.equal(204);
+    });
+  });
 });
